WIP: map rework

+++ includes/cregex.h
@@ -17,7 +17,7 @@
 
 ////////////////////////////////////////////////////////////////////////////////
 
-int regex_match(const char *pattern, char *subject);
+ssize_t regex_match(const char *pattern, char *subject);
 char *regex_replace(const char *pattern, const char *repl, char *subject);
 char **regex_capture(const char *pattern, char *subject);
 char **regex_split(char *pattern, char *subject);
